Refactor sync checks
Move checks of formats into formats/, and individual assertions into
checks/
Added an EditionBase format
