+++ lib/sync_checker/run_detailed_guides.rb
@@ -6,6 +6,6 @@ else
 end
 
 documents = Document.where(id: ids)
-dg_checks = documents.map { |doc| SyncChecker::DetailedGuideCheck.new(doc) }
+dg_checks = documents.map { |doc| SyncChecker::Formats::DetailedGuideCheck.new(doc) }
 checker = SyncChecker::SyncCheck.new(dg_checks, csv_file_path: ARGV[0])
 checker.run
